refactor(energy): move the request defaults out of the panel component
App.tsx seeded its state by importing constants from SolarPanel, so the
application state depended on a view component. Reorganising that panel -- which
the energy screen requires -- would have broken state that has nothing to do with
how the panel is laid out.

The energy defaults and the wind defaults go to lib, in two files rather than
one. Wind is a different source on a different grid with its own run kind, and
its constants living in a file named for the solar panel was the file-level form
of the coupling this undoes.

No values change. Each keeps the citation or the project-convention marking it
already carried.
